+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## v2.0.3 (1.21)
 - improved stat preview more
 - fixed bugs related enchanting and to twin blades enchanting
+- added enchantability to material view